updated user markers

+++ app/src/main/java/com/nesmelov/alexey/vkfindme/storage/OnUserUpdatedListener.java
@@ -1,6 +1,6 @@
 package com.nesmelov.alexey.vkfindme.storage;
 
 public interface OnUserUpdatedListener {
-    void onUserUpdated(final long userId, final double lat, final double lon);
-    void onUserInvisible(final long userId);
+    void onUserUpdated(final Integer userId, final double lat, final double lon);
+    void onUserInvisible(final Integer userId);
 }
